requirements.txt
Flask==3.1.0
langchain_core==0.3.21
langchain_google_genai==2.0.6
numpy==2.1.3
pydantic==2.10.3
python-dotenv==1.0.1
Requests==2.32.3
tensorflow==2.18.0
tensorflow_intel==2.18.0
